requirements.txt
boto3>=1.4.3,<2
bs4==0.0.1
cassandra-driver>=3.12.0,<4
cryptography==2.3.1
dnspython==1.15.0
dogpile.cache==0.6.7
eventlog-writer==0.4.5
flask-restful-swagger==0.19
Flask-RESTful==0.3.5
Flask==0.10.1
functional==0.4
ipaddress==1.0.22
Jinja2==2.10
jsonpath-rw==1.4.0
MarkupSafe==1.0
numpy==1.15.1
opentracing>=1.2.2,<2
opentracing-utils>=0.14,<1
prometheus-client==0.3.1
pykube==0.15.0
pymemcache>=1.4.3
pymongo==3.0.3
pymssql>=2.1.3,<3
PyMySQL==0.6.1
pysnmp==4.4.5
python-ldap==3.2.0
python-snappy==0.5.3
python-whois==0.7.1
pytz==2018.5
PyYAML>=4.2b1,<5
redis==2.10.6
requests>=2.7.0,<3
setproctitle==1.1.10
stups-tokens>=1.0.14,<2
subprocess32==3.5.2
timeperiod2==0.1
Yapsy==1.11.223
